Use cxxopts to parse command-line arguments
[Why]
Originally I added some string helpers to do the parsing. But this
is error prone and I wanna to play around with wrap files a bit, so
this is a good opportunity to do so.

[How]
Add a subproject and add cxxopts as dependency to meson.
Then replace the sting parsing with cxxopts and drop all the now
(and previously) unused functions.

Signed-off-by: Soeren Grunewald <soeren.grunewald@gmx.net>